==> WBSCodingSchool/Eniac_2/01_Data_exploration.ipynb <==
# Eniac —the e-commerce

The company has high hopes put into the possibilities that come with Data Analysis, and they are especially hopeful that your work can finally settle an ongoing debate: whether or not it’s beneficial to discount products.

* The Marketing Team Lead is convinced that offering discounts is beneficial in the long run. She believes discounts improve customer acquisition, satisfaction and retention, and allow the company to grow.
* The main investors in the Board are worried about offering aggressive discounts. They have pointed out how the company’s recent quarterly results showed an increase in orders placed, but a decrease in the total revenue. They prefer that the company positions itself in the quality segment, rather than competing to offer the lowest prices in the market.

%pip install pandas numpy

## Load necessary libraries and datasets

import pandas as pd
import numpy as np 
import math

brands_df = pd.read_csv("./Datasets/brands.csv")
orderlines_df = pd.read_csv("./Datasets/orderlines.csv")
orders_df = pd.read_csv("./Datasets/orders.csv")
products_df = pd.read_csv("./Datasets/products.csv")

## First glance and task statements

### Brands

brands_df.info(), 
brands_df.duplicated().sum()

Clear 0 duplicates/ 0 null shape(187,2)

### Orderlines

orderlines_df.info() 
orderlines_df.duplicated().sum()

orderlines_df.head()

orderlines_df.describe()

### Orderlines_df tasks
We need to:
* drop product_id - all nulls
* convert unit_price to float
* convert date from object to date


### Orders

orders_df.info() 
orders_df.duplicated().sum()

orders_df['total_paid'].isna().value_counts()

orders_df.head()

### Orders_df tasks
We need to:

* convert date from object to date
* delete all null values in total_paid columns

### Products

products_df.info() 
products_df.duplicated().sum()

products_df.describe()


products_df.sample(10)

products_df.head()

### Products_df tasks
We need to:

* drop 8746 duplicates
* solve problem with additional dots in price and promo_price columns
* convert price and promo_price
* deside what to do with in_stock columns
