==> WBSCodingSchool/Eniac_2/02_Data_cleaning.ipynb <==
# Eniac — the e-commerce

## So our task in this noutbook is:

__With Orderlines_df:__
* drop product_id - all nulls
* convert unit_price to float
* convert date from object to date

__Orders_df tasks:__
* convert date from object to date
* delete all null values in total_paid columns

__Products_df tasks:__
* drop 8746 duplicates
* solve problem with additional dots in price and promo_price columns
* convert price and promo_price
* deside what to do with in_stock columns



import os
import pandas as pd 
import numpy as np
from utils import *

brands_df = pd.read_csv("./Datasets/brands.csv")
orderlines_df = pd.read_csv("./Datasets/orderlines.csv")
orders_df = pd.read_csv("./Datasets/orders.csv")
products_df = pd.read_csv("./Datasets/products.csv")

## Orderlines_df

orderlines_df.info()

orderlines_df = orderlines_df.drop(columns="product_id")

orderlines_df.sample(10)

# as we can see, we found problems with additional dots in products_d,f so check it here
check_dots(orderlines_df, "unit_price")

# correct these numbers

to_remove_extra_dots(orderlines_df, "unit_price")

orderlines_df.describe()

As we can see, after the corrections, we received a bit of weird outliers with one negative price, and one price that abnormally high. So let`s check these prices and correct or drop them.

min_mask = (orderlines_df['unit_price'] < 0) 
max_mask = (orderlines_df['unit_price'] == orderlines_df['unit_price'].max())
min_value = orderlines_df.loc[min_mask] 
max_value = orderlines_df.loc[max_mask, :]

print(min_value, '\n**\n ', max_value)

#to check min and max price in order df and products df
## min
print(min_value)
print(products_df.loc[products_df['sku'] == min_value.iloc[0,4]])
print()
print(orders_df.loc[orders_df['order_id'] == min_value.iloc[0,1]])

As we can see this order was Cancelled, we can drop it

orderlines_df = orderlines_df.drop(min_value.index, axis=0)
orderlines_df.describe()

##max
print(max_value)
print(products_df.loc[products_df['sku'] == max_value.iloc[0,4]])
print(orders_df.loc[orders_df['order_id'] == max_value.iloc[0,1]])

The same procedure for the max value

# to normalise Date and final checks

orderlines_df["date"] = pd.to_datetime(orderlines_df["date"])


print(f"Duplicated: {orderlines_df.duplicated().sum()}")
print(f"Null values:\n{orderlines_df.isna().sum()}")

orderlines_df.info()
orderlines_df.sample(20)

## Orders_df

orders_df.info()

print(orders_df.duplicated().sum())
print(orders_df['total_paid'].isna().value_counts())

orders_df = orders_df.dropna(axis=0)
orders_df["created_date"] = pd.to_datetime(orders_df["created_date"])

orders_df.info()

## Products_df

products_df.info()

products_df = products_df.dropna()
products_df= products_df.drop_duplicates()
products_df

products_df.info()
orderlines_df.info()

check_dots(products_df, "price")
to_remove_extra_dots(products_df, "price")

orderlines_df_un = orderlines_df.groupby("sku").agg({"unit_price":"mean"})
orderlines_df_un = orderlines_df_un.reset_index()
orderlines_df_un

# from utils import price_correction #one more way to solve this problem

products_df_cp_clear_price = (
    products_df.loc[:, ['sku','name','desc', 'price',
                           'promo_price', 'in_stock', 'type']].
                           merge(
                              orderlines_df_un.loc[:, ['sku','unit_price']],
                              on='sku',
                              how='inner'
))
products_df_cp_clear_price.shape

# products_df_cp_clear_price['price_correction'] = products_df_cp_clear_price.apply(price_correction, axis=1) #one more way to solve this problem



diff_mask = products_df_cp_clear_price["price"]/products_df_cp_clear_price["unit_price"] > 2.6
products_df_cp_clear_price.loc[diff_mask]

import numpy as np

conditions = [
    products_df_cp_clear_price["unit_price"] == 0,
    products_df_cp_clear_price["price"] / products_df_cp_clear_price["unit_price"] < 2.6
]

choises = [
    products_df_cp_clear_price["price"],
    products_df_cp_clear_price["price"]
]

products_df_cp_clear_price["price_corrections"] = np.select(conditions, choises, default=products_df_cp_clear_price["price"]/10)

products_df_cp_clear_price

diff_mask_ = products_df_cp_clear_price["price_corrections"]/products_df_cp_clear_price["unit_price"] > 2.6
products_df_cp_clear_price = products_df_cp_clear_price.loc[~diff_mask_]
products_df_cp_clear_price

products_df_copy = products_df_cp_clear_price.drop(columns=["price","promo_price","unit_price"])


products_df_copy = products_df_copy.rename(columns={"price_corrections":"price"})
products_df_copy

products_df_copy.info()

local_path = "./Datasets/Clean/"

if not os.path.exists(local_path):
    os.makedirs(local_path)
    
brands_df.to_csv("./Datasets/Clean/brands.csv")
orderlines_df.to_csv("./Datasets/Clean/orderlines.csv")
orders_df.to_csv("./Datasets/Clean/orders.csv")
products_df_copy.to_csv("./Datasets/Clean/products.csv")
