==> WBSCodingSchool/Eniac_2/03_Data_quality.ipynb <==
# Eniac — the e-commerce

import os
import pandas as pd 

brands_cl = pd.read_csv('./Datasets/Clean/brands.csv')
orderlines_cl = pd.read_csv('./Datasets/Clean/orderlines.csv')
orders_cl = pd.read_csv('./Datasets/Clean/orders.csv')
products_cl = pd.read_csv('./Datasets/Clean/products.csv')

## Products

### How many products are not in the product table

products_cl

orderlines_cl.info()

products_cl.dropna(subset=['price'], inplace=True)

orderlines_cl_dropped_sku_set = set(orderlines_cl.sku)  # all SKUs which are present in orderlines
products_cl_sku_set = set(products_cl.sku)              # all SKU which are present in products

# len(orderlines_cl_dropped_sku_set), len(products_cl) #Output: (6048, 378395)

# check these masks for products which are not exict in product table
orderlines_cl_dropped_sku_mask = ~orderlines_cl.sku.isin(products_cl_sku_set)
print(len(orderlines_cl_dropped_sku_mask))

orderlines_cl_dropped_sku_mask = orderlines_cl.loc[orderlines_cl_dropped_sku_mask, :]

# drop these products
orderlines_cl_dropped = orderlines_cl.drop(orderlines_cl_dropped_sku_mask.index, axis=0)
print(len(orderlines_cl_dropped))

# products_compl = products_cl.copy()
products_cl.isna().sum()


## Orders and Orderlines

### Completed orders vs Other states

orders_cl_compl = orders_cl.loc[orders_cl['state'] == 'Completed']

orders_cl_exclud_compl = orders_cl.loc[orders_cl['state'] != 'Completed']
orders_cl_compl.shape, orders_cl_exclud_compl.shape, orders_cl.shape

print('****ORDERS**** ', orders_cl_compl.order_id.nunique(), orders_cl_compl.shape[0])
print('****ORDERS-LINES**** ', orderlines_cl_dropped.id_order.nunique(), orderlines_cl_dropped.shape[0])

### How many unique orders do we have in orders and how do these numbers correlate with orderliness

orderlines_cl_dropped

orders_cl_compl

# Unique IDs from the orders tabel
orders_cl_compl_id_set = set(orders_cl_compl.order_id)
print(len(orders_cl_compl_id_set)) # Output: 46605

# Make the set of the same numbers in ordder-lines table
orderlines_cl_same_id = orderlines_cl_dropped.id_order.isin(orders_cl_compl_id_set)
print(orderlines_cl_same_id.sum())  #Output: 61399

# invert our condition to finde all ids which are not exist in order table
orders_to_drop_ordlns_mask = orderlines_cl_dropped.loc[~orderlines_cl_same_id, :]

# drop all these rows
orderlines_cl_dropped = orderlines_cl_dropped.drop(orders_to_drop_ordlns_mask.index, axis=0 )

orderlines_compl = orderlines_cl_dropped.copy()

orderlines_compl

# the same steps for orders
orderlines_cl_id_set = set(orderlines_compl.id_order)
orders_cl_compl_same_id = orders_cl_compl.order_id.isin(orderlines_cl_id_set)

print(len(orders_cl_compl_id_set), len(orders_cl_compl_same_id)) #Output: (46605, 46605)
orders_cl_compl_same_id_mask = orders_cl_compl.loc[~orders_cl_compl_same_id, :]
print(len(orders_cl_compl_same_id_mask)) #Output: 417
orders_cl_compl = orders_cl_compl.drop(orders_cl_compl_same_id_mask.index, axis=0)
orders_compl = orders_cl_compl.copy()

orders_compl

print('****ORDERS**** ', orders_cl_compl.order_id.nunique(), orders_cl_compl.shape[0])
print('****ORDERS-LINES**** ', orderlines_cl_dropped.id_order.nunique(), orderlines_cl_dropped.shape[0])

## Revenue

### Calculation about revenue, discounts and outliers

* __Revenue__

orderlines_compl['revenue'] = orderlines_compl.product_quantity * orderlines_compl.unit_price
orderlines_compl

Total Revenue (12 month)

print(round(orderlines_compl.loc[orderlines_compl.date.between("2017-01-01 00:00:00", "2017-12-31 23:59:59")]["revenue"].sum(), 2))

Avg Revenue (1 month)

print(round(orderlines_compl.loc[orderlines_compl.date.between("2017-01-01 00:00:00", "2017-12-31 23:59:59")]["revenue"].sum()/12,2))

Avg Revenue by order

orderlines_compl.loc[orderlines_compl.date.between("2017-01-01 00:00:00", "2017-12-31 23:59:59")].groupby(["id_order"]).agg({"revenue":"sum"}).mean()

orders_cl_compl = orders_cl_compl.rename(columns = {"Unnamed: 0": "index"})
orderlines_compl = orderlines_compl.rename(columns = {"Unnamed: 0": "index"})
orders_cl_compl.set_index("index")
orderlines_compl.set_index("index")

local_path = "./Datasets/Qual/"

if not os.path.exists(local_path):
    os.makedirs(local_path)
    
brands_cl.to_csv("./Datasets/Qual/brands.csv")
orderlines_compl.to_csv("./Datasets/Qual/orderlines.csv")
orders_cl_compl.to_csv("./Datasets/Qual/orders.csv")
products_cl.to_csv("./Datasets/Qual/products.csv")
