==> WBSCodingSchool/Eniac_2/04_Data_categorisation .ipynb <==
# Eniac — the e-commerce

## Data categorisation 

import os
import pandas as pd 
from utils import *

brands_cl = pd.read_csv('./Datasets/Qual/brands.csv') 
orderlines_cl = pd.read_csv('./Datasets/Qual/orderlines.csv')
orders_cl = pd.read_csv('./Datasets/Qual/orders.csv')
products_cl = pd.read_csv('./Datasets/Qual/products.csv')

products_cl["category"] = ""

keys_dictionary = {
    'Protection, Cases, Backpacks': ['bag', 'backpack', 'case cover', 'Leather Case', "cover",
                              'Case Macbook', 'compact case', 'Travel Case', 'case with cover', 
                              'Aluminum case', 'case', 'Protective Case', 'Case adaptable', 'case', 'protec', 
                              'Protector', 'Keyboard Protector', 'Cover for', 'Protective housing', 
                              'Polycarbonate shell', 'Protective stand', 'screen protectors', 'Safety housing',
                              'Screen Saver', 'Leather Tassel'],

    'Repair service, Install kit, Extended warranty': ['Repair service', 'ifix', 'Repair', 
                                                       'Voltage Meter', 'Replacement'],

    'Input devices, styluses': ['keyboard', 'keypad', "trackpad", 'mouse', 'padlock', 'Hub', 'numpad', 'PCIe'],

    'NAS, HDD, SSD, Memory stick and cards':['Hard Drive', 'SSD', 'HDD', 'WD', 'Seagate', 
                'MX300', 'expansion kit', 'RAID', 'outer case',
                'DAS RAID', 'Nas', 'expansion module for', 'External Disk', 'external hard drive', 
                'NAS Server',  'external storage', 'Server', 'Hardware RAID'],

    'Sound, Headphones, Players':['headphone', 'speake', 'Headset', 'ipod', 'iPod Touch', 'music receiver',
                                  'handsfree'],

    'Connectivity & Adapters': ['connector', 'Adapter', ' adapter', 'Thunderbolt', ],
    'Charging Dock, Chargers, Stands, Supports':['Lightning charger', 'Lightning', 'MagSafe', 'Docking Station', 
                                                 'Lightning Dock', 'charging station', "support", "Safety support", 'Aluminum support',
                                                 'Support Retina','Mac suppor', 'Support MacBook', "car charger", 'car holder', 'support for'],
    
    'Power Banks, Batteries' : ['Power Bank', 'battery'],

    'Monitors':['Monitor.', '4K HDR', 'Display','HDMI','FHD', 'LED', 'HDR', 
                'LCD Monitor', 'Ultra High Definition', 'Samsung Monitor', 'Full HD', '4K monitor'],
}

type_dict = {
    'NAS, HDD, SSD, Memory stick and cards' : ['5397', "1364", "1433"],
    'Protection, Cases, Backpacks': ['5403'],
    'Electronic Devises': ['5404'],
    'Sound, Headphones, Players': ['5384', "5398"],
    'Apple iMac' : ["1282", "118692158", "21632158"],
    "Apple MacBook": ["2158"],
    'Connectivity & Adapters':['5395', "1325"],
    'Routers':["1334"],
    'Apple Watch Accessories': ["2449"],
    "Power Banks, Batteries": ["1515"],
    'Monitors':['1296'],
    'Charging Dock, Chargers, Stands, Supports' : ["5399", "8696"],
    "Apple Watch":["5185"],
    "Repair service, Install kit, Extended warranty": ['5406', "5407"],
    "Cables":['1230'],
    'Apple iPad': ["1714"],
    'Apple iPod':["1715"],
    "Apple iPhone": ["1716"],
    "Input devices, styluses": ["1229", "5401", "1387"] 
    }

sku_dict = {
    'Apple iMac' : ['PAC'],
    "NAS, HDD, SSD, Memory stick and cards":["SEA", "CRU", "KIN", "WDT"]
}

products_cl = products_cl.drop(columns = "Unnamed: 0.1").rename(columns = {"Unnamed: 0": "index"})

products_cl = products_cl.set_index("index")

category_defenition_type(products_cl, type_dict)
products_cl

products_cl.category.value_counts()

category_defenition_SKU(products_cl, sku_dict)
products_cl.category.value_counts()

#empty category and type masks
category = products_cl.category == ''
type_mask = products_cl.type == 'Empty'

# brand masks
PAC_mask = products_cl.sku.str.contains('pac', case=False)
APP_mask = (products_cl.sku.str.contains('APP', case=False)  | 
            products_cl.sku.str.contains('AP2', case=False))
SEA_mask = (products_cl.sku.str.contains('SEA', case=False) | 
            products_cl.sku.str.contains('CRU', case=False) | 
            products_cl.sku.str.contains('KIN', case=False) | 
            products_cl.sku.str.contains('WDT', case=False)) 
IFX_mask = products_cl.sku.str.contains('ifx', case=False)
REP_mask = products_cl.sku.str.contains('REP', case=False)
sev_mask = products_cl.sku.str.contains('SEV', case=False)

WD_mask = products_cl.sku.str.contains('wd', case=False)
OWC_mask = products_cl.sku.str.contains('OWC', case=False)
NTE_mask = products_cl.sku.str.contains('nte', case=False)

# type masks

type_101781405_mask = products_cl.type.str.contains(('101781405'), case=False)
type_1405_mask = products_cl.type.str.contains(('1405'), case=False)
type_5397_mask = products_cl.type.str.contains(('5397'), case=False)
type_5403_mask = products_cl.type.str.contains(('5403'), case=False)
type_5404_mask = products_cl.type.str.contains(('5404'), case=False)
type_1298_mask = products_cl.type.str.contains(('1298'), case=False)
type_5384_mask = products_cl.type.str.contains(('5384'), case=False)
type_1282_mask = products_cl.type.str.contains(('1282'), case=False)
type_1364_mask = products_cl.type.str.contains(('1364'), case=False)
type_5395_mask = products_cl.type.str.contains(('5395'), case=False)
type_1334_mask = products_cl.type.str.contains(('1334'), case=False)
type_2449_mask = products_cl.type.str.contains(('2449'), case=False)
type_2158_mask = products_cl.type.str.contains(('2158'), case=False)
type_1515_mask = products_cl.type.str.contains(('1515'), case=False)
type_5398_mask = products_cl.type.str.contains(('5398'), case=False)
type_5399_mask = products_cl.type.str.contains(('5399'), case=False)
type_5185_mask = products_cl.type.str.contains(('5185'), case=False)
type_1430_mask = products_cl.type.str.contains(('5406'), case=False)
type_1230_mask = products_cl.type.str.contains(('1230'), case=False)
type_1714_mask = products_cl.type.str.contains(('1714'), case=False)
type_1433_mask = products_cl.type.str.contains(('1433'), case=False)
type_1715_mask = products_cl.type.str.contains(("1715"), case=False)
type_1716_mask = products_cl.type.str.contains(("1716"), case=False)
type_5405_mask = products_cl.type.str.contains(("5407"), case=False)
type_1325_mask = products_cl.type.str.contains(("1325"), case=False)
type_8696_mask = products_cl.type.str.contains(("8696"), case=False)
type_1229_mask = products_cl.type.str.contains(("1229"), case=False)
type_1387_mask = products_cl.type.str.contains(("1387"), case=False)
type_1231_mask = products_cl.type.str.contains(("1231"), case=False)
type_1216_mask = products_cl.type.str.contains(("1216"), case=False)
type_5405_mask = products_cl.type.str.contains(("5405"), case=False)
type_9094_mask = products_cl.type.str.contains(("9094"), case=False)
type_monitor_mask = products_cl.type.str.contains(('1296'), case=False)
type_10142_mask = products_cl.type.str.contains(("10142"), case=False)

# name and description mask

type_IFIX_mask = products_cl.name.str.contains('ifix', case=False)

type_imac_mask = (products_cl.name.str.contains('Apple iMac', case=False) |
                products_cl.desc.str.contains('Apple iMac', case=False) |
                products_cl.name.str.contains('Apple iMac', case=False) |
                products_cl.desc.str.contains('Apple iMac', case=False))

protector_mask = (products_cl.name.str.contains('Protector', case=False) | 
                    products_cl.name.str.contains('Backpack', case=False) |
                    products_cl.name.str.contains('Bandolier', case=False) |
                    products_cl.name.str.contains('Protection', case=False) |
                    products_cl.name.str.contains('cover', case=False)|
                    products_cl.name.str.contains('case', case=False)|
                    products_cl.desc.str.contains('Protector', case=False) | 
                    products_cl.desc.str.contains('Backpack', case=False) |
                    products_cl.desc.str.contains('Bandolier', case=False) |
                    products_cl.desc.str.contains('Protection', case=False) |
                    products_cl.desc.str.contains('cover', case=False)|
                    products_cl.desc.str.contains('case', case=False) | 
                    products_cl.name.str.contains('housing', case=False)|
                    products_cl.desc.str.contains('housing', case=False) | 
                    products_cl.name.str.contains('Tassel', case=False)|
                    products_cl.desc.str.contains('Tassel', case=False))

monitor_mask = (products_cl.name.str.contains('Monitor', case=False) |
                products_cl.desc.str.contains('Monitor', case=False))
products_cl.loc[category &  monitor_mask, 'category'] = "Monitors"
camera_mask = (products_cl.name.str.contains('camera', case=False)|
               products_cl.name.str.contains('secur', case=False)|
                products_cl.desc.str.contains('night', case=False)|
                products_cl.desc.str.contains('extend', case=False)
                )

strap_mask = (products_cl.desc.str.contains('strap | armband', case=False)|
                products_cl.name.str.contains('strap | armband', case=False)|
                products_cl.desc.str.contains('bands', case=False)|
                products_cl.name.str.contains('bands', case=False)|
                products_cl.desc.str.contains('bracelet', case=False)|
                products_cl.name.str.contains('bracelet', case=False))


appl_watch_mask = (products_cl.desc.str.contains('Apple Watch', case=False)|
                products_cl.name.str.contains('Apple Watch', case=False))

products_cl.loc[category & (type_5405_mask | strap_mask) & appl_watch_mask, 'category'] = "Apple Watch Accessories"

port_mask  = (products_cl.desc.str.contains('display port', case=False)|
                products_cl.name.str.contains('display port', case=False))

cable_mask = (products_cl.desc.str.contains('cable', case=False)|
                products_cl.name.str.contains('cable', case=False)
                )

headphoenes = (products_cl.desc.str.contains('headphones', case=False) |
                products_cl.name.str.contains('headphones', case=False))

microph_mask = (products_cl.name.str.contains("microphone | speaker | handsfree | player", case=False) | 
                products_cl.desc.str.contains("microphone | speaker | handsfree | player", case=False) )

products_cl.loc[category & (headphoenes | microph_mask), 'category'] = "Sound, Headphones, Players"

category = products_cl.category == ''
type_mask = products_cl.type == 'Empty'
 
appmac_mask = (products_cl.name.str.contains('Apple iMac', case=False) |
                products_cl.desc.str.contains('PC'))

nas_HD_mask =  (products_cl.name.str.contains("Hard Drive | drive | nas | " \
                        "Ethernet | raid | SSD | nas server | Hard Thunderbolt" \
                        "| hard disk | external disk | Disk 4TB", case=False) |
                    products_cl.desc.str.contains("Hard Drive | drive | nas | " \
                        "Ethernet | raid | SSD | nas server | Hard Thunderbolt" \
                        "| hard disk | external disk | Disk 4TB", case=False))

products_cl.category.value_counts()

not_include_mask = (products_cl.name.str.contains("Satechi Sonic Dual Conical Mac ", case=False)|
                    products_cl.desc.str.contains("tablet | card | protect |microphone | display | pointer ", case=False)|
                   
                    products_cl.name.str.contains("hub | adapter", case=False) |
                    products_cl.desc.str.contains("hub | adapter", case=False) |
                    products_cl.name.str.contains("license |software |  Adobe InDesign", case=False) |
                    products_cl.desc.str.contains("license |software |  Adobe InDesign", case=False) 
                    )

# CASES & Protection Mask 363

category = products_cl.category == ''

type_other_cases_mask = (products_cl.name.str.contains('case | housing | bag | backpack | case cover '
                            '| Leather Case | cover | Case Macbook | compact case | Travel Case | ' \
                            'case with cover | Aluminum case | case | Protective Case | Case adaptable |' \
                            ' case | protec | Protector | Keyboard Protector | Cover for | Protective housing |' \
                            'Polycarbonate shell | Protective stand | screen protectors | Safety housing |' \
                            'Screen Saver | Leather Tassel | Defend | folio | bag | bags |backpack  for', case=False) |
                        products_cl.desc.str.contains(('case | housing | bag | backpack | case cover'
                            ' | Leather Case | cover | Case Macbook | compact case | Travel Case |'
                            ' case with cover | Aluminum case | case | Protective Case | Case adaptable | '
                            'case | protec | Protector | Keyboard Protector | Cover for | Protective housing |' \
                            'Polycarbonate shell | Protective stand | screen protectors | Safety housing |' \
                            'Screen Saver | Leather Tassel | Defend | folio | bag | bags |backpack  for'), case=False)) 

type_OWC_cases_mask = products_cl.sku.str.contains("owc", case=False) # memory, doc
type_BTC_cases_mask = products_cl.sku.str.contains("BTC", case=False) 
type_THU_cases_mask = products_cl.sku.str.contains("thu", case=False) 
type_hic_cases_mask = products_cl.sku.str.contains("hic", case=False) 
type_mos_cases_mask = products_cl.sku.str.contains("mos", case=False) 
type_mdt_cases_mask = products_cl.sku.str.contains("mdt", case=False) 
type_lif_cases_mask = products_cl.sku.str.contains("lif", case=False) # armband support Belt Clip  accessory adapter

products_cl.loc[(type_other_cases_mask) & category, 'category'] = 'Protection, Cases, Backpacks'
products_cl.loc[(type_BTC_cases_mask) & category, 'category'] = 'Protection, Cases, Backpacks'
# print(products_cl.loc[category & type_other_cases_mask].shape)
# products_cl.loc[category & type_other_cases_mask].sample(1)  

products_cl.category.value_counts()

# Repair service, Install kit, Extended warranty #154


repair_mask = (products_cl.name.str.contains('repair | service | calibrator', case=False) |
                products_cl.desc.str.contains('repair | service | calibrator', case=False)|
                products_cl.desc.str.contains('including parts', case=False)
                )#

 

# products_cl.loc[(IFX_mask | repair_mask) & category, 'category'] = "Repair service, Install kit, Extended warranty"# onlz 132 row
products_cl.loc[(IFX_mask | REP_mask | sev_mask | repair_mask), 'category'] = "Repair service, Install kit, Extended warranty"


print(products_cl.loc[(IFX_mask | REP_mask | sev_mask | repair_mask)].shape)
products_cl.loc[(IFX_mask | REP_mask | sev_mask | repair_mask) ].sample(6) 
products_cl.category.value_counts()
# 'Repair service', 'ifix', 'Repair', 'Voltage Meter', 'Replacement'

# NAS, HDD, SSD, Memory stick and cards 176 row

memo_mask = (products_cl.name.str.contains('memory | pen drive | pendrive | ssd | nas | hard drive | flash drive |' \
            'shdc | memory | hdd | raid | SSD | HDD | WD | Seagate | outer case | DAS RAID | expansion module for | ' \
            'External Disk | external hard drive | NAS server | storage | server | hardware raid | drive |TB portable', case=False) |
              products_cl.desc.str.contains('memory | pen drive | pendrive | ssd | nas | hard drive | flash drive | ' \
              'shdc | memory | hdd | raid | SSD | HDD | WD | Seagate | outer case | DAS RAID | expansion module for | ' \
              'External Disk | external hard drive | NAS server | storage | server | hardware raid | drive |TB portable', case=False))
memo_mask_ = products_cl.sku.str.contains('SEA | CRU | KIN | WDT', case=False)

warranty_mask = (products_cl.desc.str.contains('warranty | plan | care | extended warranty', case=False) |  # Did not use 15 rows  with 1231_mask!!
                 products_cl.name.str.contains('warranty | plan | care | extended warranty', case=False) 
                 )

print(products_cl.loc[category & (memo_mask | memo_mask_) & ~ APP_mask].shape)
products_cl.loc[category & (memo_mask | memo_mask_)& ~ APP_mask & ~warranty_mask].sample(len(products_cl.loc[category & (memo_mask | memo_mask_)& ~ APP_mask &~warranty_mask]))  


products_cl.loc[category & (memo_mask | memo_mask_)& ~ APP_mask & ~warranty_mask, 'category'] = "NAS, HDD, SSD, Memory stick and cards"
products_cl.category.value_counts()

# Apple MacBook 242 rows

type_macbook_mask = (products_cl.name.str.contains('Apple Macbook', case=False)
                     | products_cl.desc.str.contains('Apple Macbook', case=False) |
                     products_cl.name.str.contains('Macbook', case=False)
                     | products_cl.desc.str.contains('Macbook', case=False) |
                     products_cl.name.str.contains('laptop', case=False)
                     | products_cl.desc.str.contains('laptop', case=False))

charger_mask = (products_cl.name.str.contains('MagSafe', case=False)|
                products_cl.desc.str.contains('MagSafe', case=False)|
                products_cl.desc.str.contains('charg', case=False) |
                products_cl.name.str.contains('charg', case=False) |
                products_cl.desc.str.contains('holder', case=False) |
                products_cl.name.str.contains('support | Mount', case=False) |
                products_cl.desc.str.contains('support | Mount', case=False)
                )
battery_mask = (products_cl.desc.str.contains('battery | mAh | power bank', case=False)|
                products_cl.name.str.contains('battery | mAh | power bank', case=False)|
                products_cl.desc.str.contains('W ')| products_cl.name.str.contains('W '))

print(products_cl.loc[category & type_macbook_mask & APP_mask & ~(charger_mask | battery_mask)].shape)
products_cl.loc[category & type_macbook_mask & APP_mask & ~(charger_mask | battery_mask)].sample(20)

products_cl.loc[category & type_macbook_mask & APP_mask & ~(charger_mask | battery_mask), 'category'] = "Apple MacBook"
products_cl.category.value_counts()

# Apple iMac 133

category = products_cl.category == ''
type_imac_mask = (products_cl.name.str.contains('Apple iMac | desktop | PC | desktop computer ', case=False) |
                products_cl.desc.str.contains('Apple iMac | desktop | PC | desktop computer ', case=False))

software_mask = (products_cl.desc.str.contains('Software license', case=False) |
                  products_cl.name.str.contains('Software license', case=False) |
                  products_cl.desc.str.contains('software | microsoft', case=False) |
                  products_cl.name.str.contains('software | microsoft', case=False) |
                  products_cl.desc.str.contains('licenses', case=False) |
                  products_cl.name.str.contains('licenses', case=False) |
                  products_cl.desc.str.contains('Window') |
                  products_cl.name.str.contains('Window'))

type_connector_mask = (products_cl.name.str.contains(('adapter | connector | adap'), case=False)|
                       products_cl.desc.str.contains(('adapter | connector | adap'), case=False))


print(products_cl.loc[category & type_imac_mask & ~(software_mask | type_connector_mask | camera_mask)].shape)
products_cl.loc[category & type_imac_mask & ~(software_mask | type_connector_mask | camera_mask)].sample(20)

products_cl.loc[category & type_imac_mask & ~(software_mask | type_connector_mask | camera_mask), 'category'] = "Apple iMac"
products_cl.category.value_counts()

#Apple iPhone 42

type_iphone_mask = (products_cl.name.str.contains('Apple iPhone', case=False) |
                    products_cl.desc.str.contains('Apple iPhone', case=False))


print(products_cl.loc[category & type_iphone_mask].shape)
# products_cl.loc[category & type_iphone_mask].sample(20)

products_cl.loc[category & type_iphone_mask, 'category'] = "Apple iPhone"
products_cl.category.value_counts()

# Apple iPad 12 row

type_ipad_mask = (products_cl.name.str.contains('Apple iPad', case=False) |
                  products_cl.desc.str.contains('Apple iPad', case=False)
                  )
 
print(products_cl.loc[category & type_ipad_mask & ~protector_mask].shape)
# products_cl.loc[category & type_ipad_mask & ~protector_mask].sample(12)

products_cl.loc[category & type_ipad_mask & ~protector_mask, 'category'] = "Apple iPad"
products_cl.category.value_counts()

# Connectivity & Adapters 54 rows

type_connector_mask = (products_cl.name.str.contains(('adapter | connector | adap'), case=False)|
                       products_cl.desc.str.contains(('adapter | connector | adap'), case=False))
 
print(products_cl.loc[category & type_connector_mask & ~(protector_mask | charger_mask)].shape)
products_cl.loc[category & type_connector_mask & ~(protector_mask | charger_mask)].sample(12)

products_cl.loc[category & type_connector_mask & ~(protector_mask | charger_mask), 'category'] = "Connectivity & Adapters"
products_cl.category.value_counts()

# Input devices, styluses 45
category = products_cl.category == ''
type_pens_mask = (products_cl.name.str.contains(('mouse | trackpad | keyapd | keyboard | stylus | pointer | pen'), case=False)|
                    products_cl.desc.str.contains(('mouse | trackpad | keyapd | keyboard stylus | pointer | pen'), case=False))

tablet_mask = (products_cl.name.str.contains('tablet', case=False) | 
               products_cl.desc.str.contains('tablet', case=False))

print(products_cl.loc[category & type_pens_mask & ~(tablet_mask)].shape)
products_cl.loc[category & type_pens_mask & ~(tablet_mask)].sample(20)

products_cl.loc[category & type_pens_mask &~tablet_mask & ~charger_mask, 'category'] = "Input devices, styluses"
products_cl.category.value_counts()

# Tablets 75 rows

tablet_mask = (products_cl.name.str.contains('tablet', case=False) | 
               products_cl.desc.str.contains('tablet', case=False))


charger_mask = (products_cl.name.str.contains('MagSafe', case=False)|
                products_cl.desc.str.contains('MagSafe', case=False)|
                products_cl.desc.str.contains('charg', case=False) |
                products_cl.name.str.contains('charg', case=False) |
                products_cl.desc.str.contains('holder', case=False) |
                products_cl.name.str.contains('support | Mount', case=False) |
                products_cl.desc.str.contains('support | Mount', case=False)
                )

discount_mask = products_cl.desc.str.contains("discount", case=False)

print(products_cl.loc[category & tablet_mask & ~ discount_mask].shape)
products_cl.loc[category & tablet_mask ].sample(20)


products_cl.loc[category & tablet_mask & ~ discount_mask, 'category'] = "Tablets"
products_cl.category.value_counts()

# Monitors 53 row
monitor_mask = (products_cl.name.str.contains('Monitor | 4K HDR | Display | HDMI | FHD | LED | HDR' \
                'LCD Monitor | Ultra High Definition | Samsung Monitor | Full HD | 4K monitor', case=False) |
                products_cl.desc.str.contains('Monitor | 4K HDR | Display | HDMI | FHD | LED | HDR' \
                'LCD Monitor | Ultra High Definition | Samsung Monitor | Full HD | 4K monitor', case=False))

watch_mask = (products_cl.name.str.contains('smartwatch', case=False) |
                products_cl.desc.str.contains('smartwatch', case=False))
dockst_mask = (products_cl.name.str.contains('Docking station | Dock USB-C multiple | Dock with | dock | docking | USB Hub | hub thunder', case=False)|
                products_cl.desc.str.contains('Docking station | Dock USB-C multiple | Dock with | dock | docking | USB Hub | hub thunder', case=False) 
                )#USB Hub



print(products_cl.loc[category & monitor_mask & ~(camera_mask | cable_mask | battery_mask | watch_mask | strap_mask | port_mask | dockst_mask | type_connector_mask )].shape)
# products_cl.loc[category & monitor_mask &~(camera_mask | cable_mask | battery_mask | watch_mask | strap_mask | port_mask | dockst_mask | type_connector_mask)].sample(11)

products_cl.loc[category & monitor_mask & ~(camera_mask | cable_mask | battery_mask | watch_mask 
                                            | strap_mask | port_mask | dockst_mask | type_connector_mask ), 'category'] = "Monitors"
products_cl.category.value_counts()

# "Cameras, Security system" 46
camera_mask = (products_cl.name.str.contains('camera | Security Cameras | smart security cameras' \
                '| Safety System | cameras | secur | night | extend', case=False)|
               products_cl.desc.str.contains('camera | Security Cameras | smart security cameras' \
                '| Safety System | cameras | secur | night | extend', case=False)
                )

warranty_mask = (products_cl.desc.str.contains('warranty | plan | care | extended warranty', case=False) |  # Did not use 15 rows  with 1231_mask!!
                 products_cl.name.str.contains('warranty | plan | care | extended warranty  | Security Cameras | smart security cameras ' \
                 ' | Safety System | cameras', case=False) 
                 )

print(products_cl.loc[category & (type_9094_mask & camera_mask)].shape)
products_cl.loc[category & (type_9094_mask & camera_mask)].sample(len(
                        products_cl.loc[category & (type_9094_mask & camera_mask)]))


products_cl.loc[category & (type_9094_mask & camera_mask) , 'category'] = "Cameras, Security system"
products_cl.category.value_counts()

#Charging Dock, Chargers, Stands, Supports 128

charger_mask = (products_cl.name.str.contains('MagSafe | charg | holder | support | mount', case=False)|
                products_cl.desc.str.contains('MagSafe | charg | holder | support | mount', case=False)
                )

dockst_mask = (products_cl.name.str.contains('Docking station | Dock USB-C multiple | Dock with | dock | docking | USB Hub | hub thunder', case=False)|
                products_cl.desc.str.contains('Docking station | Dock USB-C multiple | Dock with | dock | docking | USB Hub | hub thunder', case=False) 
                )#USB Hub

print(products_cl.loc[category &  (type_1216_mask | charger_mask | dockst_mask) & ~(protector_mask)].shape)
products_cl.loc[category &  (type_1216_mask | charger_mask | dockst_mask) & ~(protector_mask)].sample(len(products_cl.loc[category &  (type_1216_mask | charger_mask | dockst_mask) & ~(protector_mask)]))

products_cl.loc[category &  (type_1216_mask | charger_mask | dockst_mask) & ~(protector_mask), 'category'] = "Charging Dock, Chargers, Stands, Supports"
products_cl.category.value_counts()

# Batteries 45

category = products_cl.category == ''
battery_mask = (products_cl.desc.str.contains('battery | mAh | power bank', case=False)|
                products_cl.name.str.contains('battery | mAh | power bank', case=False)|
                products_cl.desc.str.contains('W ')| products_cl.name.str.contains('W '))



print(products_cl.loc[category & (battery_mask | type_10142_mask)].shape)
products_cl.loc[category &  (battery_mask | type_10142_mask)].sample(11)

products_cl.loc[category & (battery_mask | type_10142_mask), 'category'] = "Batteries"
products_cl.category.value_counts()

# Apple Watch Accessories 50
category = products_cl.category == ''
strap_mask = (products_cl.desc.str.contains('strap | armband', case=False)|
                products_cl.name.str.contains('strap | armband', case=False)|
                products_cl.desc.str.contains('bands', case=False)|
                products_cl.name.str.contains('bands', case=False)|
                products_cl.desc.str.contains('bracelet', case=False)|
                products_cl.name.str.contains('bracelet', case=False))

print(products_cl.loc[category & (strap_mask | type_5405_mask)].shape)
products_cl.loc[category &  (strap_mask | type_5405_mask)].sample(11)

products_cl.loc[category & (type_5405_mask | strap_mask) & appl_watch_mask, 'category'] = "Apple Watch Accessories"
products_cl.category.value_counts()

# Sound, Headphones, Players 34
category = products_cl.category == ''
headphoenes_mask = (products_cl.desc.str.contains('headphones | microphone | speaker | handsfree | player', case=False) |
                products_cl.name.str.contains('headphones | microphone | speaker | handsfree | player', case=False))

print(products_cl.loc[category & headphoenes_mask].shape)
# products_cl.loc[category &  headphoenes_mask].sample(11)

products_cl.loc[category & (headphoenes_mask), 'category'] = "Sound, Headphones, Players"
products_cl.category.value_counts()

# Software, Licenses 20

software_mask = (products_cl.desc.str.contains('Software license | software | microsoft | licenses | window', case=False) |
                  products_cl.name.str.contains('Software license | software | microsoft | licenses | window', case=False))

ado_mask = (products_cl.sku.str.contains('ADO', case=False))
SOF_mask = (products_cl.sku.str.contains('SOF', case=False))

type_1416_mask = products_cl.type.str.contains('1416', case=False)

print(products_cl.loc[category & (type_1416_mask | software_mask)].shape)
products_cl.loc[category & (type_1416_mask | software_mask)].sample(10)


products_cl.loc[category & (type_1416_mask | software_mask), 'category'] = "Software, Licenses"
products_cl.category.value_counts()

# "Cables" 26

cable_mask = (products_cl.desc.str.contains('cable', case=False)|
                products_cl.name.str.contains('cable', case=False)
                )

print(products_cl.loc[category & cable_mask].shape)
products_cl.loc[category & cable_mask].sample(10)


products_cl.loc[category & cable_mask, 'category'] = "Cables"
products_cl.category.value_counts()

# Armbands, Bracelets, Belts 31
print(products_cl.loc[category & type_5405_mask].shape)
products_cl.loc[category & type_5405_mask].sample(10)


products_cl.loc[category & type_5405_mask, 'category'] = "Armbands, Bracelets, Belts "
products_cl.category.value_counts()

# "Repair service, Install kit, Extended warranty"11

warranty_mask = (products_cl.desc.str.contains('warranty | plan | care | extended warranty', case=False) |  # Did not use 15 rows  with 1231_mask!!
                 products_cl.name.str.contains('warranty | plan | care | extended warranty', case=False) 
                 )


print(products_cl.loc[category & warranty_mask].shape)
products_cl.loc[category &  warranty_mask].sample(3)

products_cl.loc[category & APP_mask & warranty_mask, 'category'] ="Repair service, Install kit, Extended warranty"
products_cl.category.value_counts()

# Network routers, Network equipment 21

# core_mask = products_qu.category.str.contains(" core ", case=False)
dlk_mask = products_cl.sku.str.contains("dlk", case=False)

router_mask = (products_cl.name.str.contains("Ethernet | Wireless | network | internet | WiFi | Access Point", case=False) |
                products_cl.desc.str.contains("Ethernet | Wireless | network | internet | WiFi | Access Point", case=False))

print(products_cl.loc[category & router_mask & ~(camera_mask & memo_mask & memo_mask_ & dockst_mask) ].shape)
products_cl.loc[category & router_mask & ~(camera_mask & memo_mask & 
                                           memo_mask_ & dockst_mask)].sample(len(products_cl.loc[category & router_mask &
                                                                                                  ~(camera_mask & memo_mask & memo_mask_ & dockst_mask)]))

products_cl.loc[category & router_mask & ~(camera_mask & memo_mask & memo_mask_ & dockst_mask) , 'category'] ="Network routers, Network equipment"
products_cl.category.value_counts()

# additional NAS, HDD, SSD, Memory stick and cards 4
hard_dr_mask = (products_cl.name.str.contains("Hard Disk | hard drive | eSata card 6Gb")|
                products_cl.desc.str.contains("Hard Disk | hard drive | eSata card 6Gb"))

type_1276_mask = products_cl.type.str.contains("1276")

print(products_cl.loc[category & type_1276_mask].shape)
products_cl.loc[category & type_1276_mask].sample(len(products_cl.loc[category & type_1276_mask]))

products_cl.loc[category & type_1276_mask, 'category'] ="NAS, HDD, SSD, Memory stick and cards"
products_cl.category.value_counts()

# additional iMac 15
APP1213_mask = products_cl.sku.str.contains("APP1213")

print(products_cl.loc[category & APP_mask & ~APP1213_mask].shape)
products_cl.loc[category & APP_mask & ~APP1213_mask].sample(len(products_cl.loc[category & APP_mask & ~APP1213_mask]))

products_cl.loc[category & APP_mask & ~APP1213_mask, 'category'] = "Apple iMac"
products_cl.category.value_counts()

# additional iMac 15
APP1213_mask = products_cl.sku.str.contains("APP1213")

print(products_cl.loc[category & APP_mask & ~APP1213_mask].shape)
products_cl.loc[category & APP_mask & ~APP1213_mask].sample(len(products_cl.loc[category & APP_mask & ~APP1213_mask]))

products_cl.loc[category & APP_mask & ~APP1213_mask, 'category'] = "Apple iMac"
products_cl.category.value_counts()

# drop other into "Electronic dev." 133
eldev_mask = products_cl.type == ("Electronic Devises")

print(products_cl.loc[category].shape)
products_cl.loc[category].sample(26)

products_cl.loc[category, 'category'] ="Electronic Devises"

products_cl.category.value_counts()

# Drones

products_cl.loc[products_cl.category == "Electronic Devises"].sort_values("price")

eldev_mask = products_cl.category == ("Electronic Devises")

drone_mask = (products_cl.name.str.contains("drone", case=False)|
                products_cl.desc.str.contains("drone", case=False))


print(products_cl.loc[drone_mask].shape)
products_cl.loc[drone_mask].sample(len(products_cl.loc[drone_mask]))

products_cl.loc[drone_mask, 'category'] ="Drones"

products_cl.category.value_counts()

products_cl.to_csv(".\Datasets\Qual\products.csv")
